+++ include/kernels/expect_count.cuh
@@ -25,9 +25,9 @@
 __global__
 #endif
 
-void kernel_expect_count(float* count, float* mu, float* beta, const uint32_t* sequence, uint32_t* pretermination_lookuptable, 
-
-                        uint32_t* grammar_index, uint32_t* grammar_table, float* alpha, 
+void kernel_expect_count(long double* count, long double* mu, long double* beta, 
+    const uint32_t* sequence, uint32_t* pretermination_lookuptable, 
+                        uint32_t* grammar_index, uint32_t* grammar_table, long double* alpha, 
                         int sequence_length, int n_syms, int N, int T, int MS, int n_grammars
                         #ifdef DEBUG_INSIDE_ALGORITHM
                         , pcfg* grammar
