+++ src/algorithms/alg_inside_outside_main.cpp
@@ -1,7 +1,10 @@
 #include "algorithms/alg_inside_outside_main.h"
 
-float* outside_algorithm(float* mu, float* beta, const uint32_t* sequence, uint32_t* pretermination_lookuptable, 
-                        uint32_t* grammar_index, uint32_t* grammar_table, float* alpha, 
+long double* outside_algorithm(long double* mu, long double* beta, const uint32_t* sequence, 
+                        uint32_t* pretermination_lookuptable, 
+                        uint32_t* grammar_index, 
+                        uint32_t* grammar_table, 
+                        long double* alpha, 
                         int sequence_length, int n_syms, int N, int T, int MS, int n_grammars,
                         std::vector<std::tuple<uint32_t, uint32_t>> inside_order_1_rule_iteration_path
                         #ifdef DEBUG_INSIDE_ALGORITHM
@@ -16,8 +19,11 @@ float* outside_algorithm(float* mu, float* beta, const uint32_t* sequence, uint3
     return beta;
 }
 
-float* em_algorithm_calculate_expection_count(float* count, float* mu, float* beta, const uint32_t* sequence, uint32_t* pretermination_lookuptable, 
-                        uint32_t* grammar_index, uint32_t* grammar_table, float* alpha, 
+long double* em_algorithm_calculate_expection_count(long double* count, long double* mu, long double* beta, const uint32_t* sequence, 
+                        uint32_t* pretermination_lookuptable, 
+                        uint32_t* grammar_index, 
+                        uint32_t* grammar_table, 
+                        long double* alpha, 
                         int sequence_length, int n_syms, int N, int T, int MS, int n_grammars
                         #ifdef DEBUG_INSIDE_ALGORITHM
                         ,pcfg* grammar
@@ -35,8 +41,11 @@ float* em_algorithm_calculate_expection_count(float* count, float* mu, float* be
     return count;
 }
 
-float* inside_algorithm(const uint32_t* sequence, uint32_t* pretermination_lookuptable, 
-                        uint32_t* grammar_index, uint32_t* grammar_table, float* alpha, 
+long double* inside_algorithm(const uint32_t* sequence, 
+                        uint32_t* pretermination_lookuptable, 
+                        uint32_t* grammar_index, 
+                        uint32_t* grammar_table, 
+                        long double* alpha, 
                         int sequence_length, int n_syms, int N, int T, int MS, int n_grammars,
                         std::vector<std::tuple<uint32_t, uint32_t>> inside_order_1_rule_iteration_path, pcfg* grammar = nullptr){
